refactor(search): Update search component exports to follow standard default export format
